requirements.txt

django
djangorestframework
rest_framework_simplejwt(deprecated)
djangorestframework-simplejwt
django-cors-headers
